+++ lib/agilibox/active_model_custom_error_messages.rb
@@ -1,0 +1,8 @@
+module Agilibox::ActiveModelCustomErrorMessages
+  def full_message(attribute, message)
+    return message[1..-1] if message[0] == "^"
+    super(attribute, message)
+  end
+end
+
+ActiveModel::Errors.send(:prepend, Agilibox::ActiveModelCustomErrorMessages)